+++ src/pages/nft/nft.tsx
@@ -90,6 +90,8 @@ import ReceiveDetail from '../assets/components/ReceiveDetail';
 import { croNftApi } from '../../service/rpc/NftApi';
 import { ExternalNftMetadataResponse } from '../../service/rpc/models/nftApi.models';
 import { Promise } from "bluebird";
+import { useLedgerStatus } from '../../hooks/useLedgerStatus';
+import { ledgerNotification } from '../../components/LedgerNotification/LedgerNotification';
 
 const { Header, Content, Footer, Sider } = Layout;
 const { TabPane } = Tabs;
@@ -209,6 +211,7 @@ const FormMintNft = () => {
   const [decryptedPhrase, setDecryptedPhrase] = useState('');
   const [broadcastResult, setBroadcastResult] = useState<BroadCastResult>({});
   const [errorMessages, setErrorMessages] = useState([]);
+  const { isLedgerConnected } = useLedgerStatus({ asset: walletAsset });
 
   const analyticsService = new AnalyticsService(currentSession);
 
@@ -280,6 +283,9 @@ const FormMintNft = () => {
 
   const showPasswordInput = () => {
     if (decryptedPhrase || currentSession.wallet.walletType === LEDGER_WALLET_TYPE) {
+      if (!isLedgerConnected) {
+        ledgerNotification(currentSession.wallet, walletAsset!);
+      }
       showConfirmationModal();
     } else {
       setInputPasswordVisible(true);
@@ -708,7 +714,10 @@ const FormMintNft = () => {
               loading={confirmLoading}
               disabled={
                 (isDenomIdIssued && !isDenomIdOwner) ||
-                new Big(multiplyFee(networkFee, !isDenomIdIssued ? 2 : 1)).gt(walletAsset.balance)
+                new Big(multiplyFee(networkFee, !isDenomIdIssued ? 2 : 1)).gt(
+                  walletAsset.balance,
+                ) ||
+                (!isLedgerConnected && currentSession.wallet.walletType === LEDGER_WALLET_TYPE)
               }
             >
               {t('general.confirm')}
@@ -954,6 +963,7 @@ const NftPage = () => {
   const [broadcastResult, setBroadcastResult] = useState<BroadCastResult>({});
   const [errorMessages, setErrorMessages] = useState([]);
   const [decryptedPhrase, setDecryptedPhrase] = useState('');
+  const { isLedgerConnected } = useLedgerStatus({ asset: walletAsset });
 
   const didMountRef = useRef(false);
 
@@ -1119,6 +1129,9 @@ const NftPage = () => {
 
   const showPasswordInput = () => {
     if (decryptedPhrase || currentSession.wallet.walletType === LEDGER_WALLET_TYPE) {
+      if (!isLedgerConnected) {
+        ledgerNotification(currentSession.wallet, walletAsset!);
+      }
       showConfirmationModal();
     } else {
       setInputPasswordVisible(true);
@@ -1637,6 +1650,10 @@ const NftPage = () => {
                       key="submit"
                       type="primary"
                       onClick={onConfirmTransfer}
+                      disabled={
+                        !isLedgerConnected &&
+                        currentSession.wallet.walletType === LEDGER_WALLET_TYPE
+                      }
                       loading={confirmLoading}
                     >
                       {t('nft.modal2.button1')}